refractor: organize the utils folder and add necessary stuff
