+++ chatbot-core/backend/app/databases/minio.py
@@ -7,9 +7,9 @@ from minio.error import S3Error
 
 from app.databases.base import BaseConnector
 from app.settings import Secrets
-from app.utils.logger import LoggerFactory
+from app.utils.api.helpers import get_logger
 
-logger = LoggerFactory().get_logger(__name__)
+logger = get_logger(__name__)
 
 
 class MinioConnector(BaseConnector[Minio]):
