+++ chatbot-core/backend/app/databases/mssql.py
@@ -8,9 +8,9 @@ from sqlalchemy.orm import sessionmaker
 from app.databases.base import BaseConnector
 from app.settings import Constants
 from app.settings import Secrets
-from app.utils.logger import LoggerFactory
+from app.utils.api.helpers import get_logger
 
-logger = LoggerFactory().get_logger(__name__)
+logger = get_logger(__name__)
 
 
 class MSSQLConnector(BaseConnector[Engine]):
