+++ chatbot-core/backend/app/databases/qdrant.py
@@ -5,9 +5,9 @@ from qdrant_client import QdrantClient
 from app.databases.base import BaseConnector
 from app.settings import Constants
 from app.settings import Secrets
-from app.utils.logger import LoggerFactory
+from app.utils.api.helpers import get_logger
 
-logger = LoggerFactory().get_logger(__name__)
+logger = get_logger(__name__)
 
 
 class QdrantConnector(BaseConnector[QdrantClient]):
