+++ chatbot-core/backend/app/repositories/chat.py
@@ -8,11 +8,11 @@ from sqlalchemy.sql import and_
 from app.models import ChatMessage
 from app.models import ChatSession
 from app.repositories.base import BaseRepository
-from app.utils.api_response import APIError
-from app.utils.error_handler import ErrorCodesMappingNumber
-from app.utils.logger import LoggerFactory
+from app.utils.api.api_response import APIError
+from app.utils.api.error_handler import ErrorCodesMappingNumber
+from app.utils.api.helpers import get_logger
 
-logger = LoggerFactory().get_logger(__name__)
+logger = get_logger(__name__)
 
 
 class ChatRepository(BaseRepository):
