+++ chatbot-core/backend/app/repositories/document.py
@@ -4,11 +4,11 @@ from sqlalchemy.orm import Session
 
 from app.models.document import DocumentMetadata
 from app.repositories.base import BaseRepository
-from app.utils.api_response import APIError
-from app.utils.error_handler import ErrorCodesMappingNumber
-from app.utils.logger import LoggerFactory
+from app.utils.api.api_response import APIError
+from app.utils.api.error_handler import ErrorCodesMappingNumber
+from app.utils.api.helpers import get_logger
 
-logger = LoggerFactory().get_logger(__name__)
+logger = get_logger(__name__)
 
 
 class DocumentRepository(BaseRepository):
