+++ chatbot-core/backend/app/routers/base.py
@@ -1,8 +1,8 @@
 from fastapi import APIRouter
 from fastapi import status
 
-from app.utils.api_response import APIResponse
-from app.utils.api_response import BackendAPIResponse
+from app.utils.api.api_response import APIResponse
+from app.utils.api.api_response import BackendAPIResponse
 
 router = APIRouter(tags=["base"])
