--- chatbot-core/backend/app/utils/accents_handler.py
@@ -1,15 +1,0 @@
-S1 = "ÀÁÂÃÈÉÊÌÍÒÓÔÕÙÚÝàáâãèéêìíòóôõùúýĂăĐđĨĩŨũƠơƯưẠạẢảẤấẦầẨẩẪẫẬậẮắẰằẲẳẴẵẶặẸẹẺẻẼẽẾếỀềỂểỄễỆệỈỉỊịỌọỎỏỐốỒồỔổỖỗỘộỚớỜờỞởỠỡỢợỤụỦủỨứỪừỬửỮữỰựỲỳỴỵỶỷỸỹ"
-S0 = "AAAAEEEIIOOOOUUYaaaaeeeiioooouuyAaDdIiUuOoUuAaAaAaAaAaAaAaAaAaAaAaAaEeEeEeEeEeEeEeEeIiIiOoOoOoOoOoOoOoOoOoOoOoOoUuUuUuUuUuUuUuYyYyYyYy"
-
-
-def remove_vietnamese_accents(input_str: str) -> str:
-    """
-    Remove Vietnamese accents from a string.
-
-    Args:
-        input_str (str): Input string with accents.
-
-    Returns:
-        str: Output string without accents.
-    """
-    return "".join([S0[S1.index(c)] if c in S1 else c for c in input_str])
